Chore: Refactor the state migration logic from the state sync
